+++ src/convtools/__init__.py
@@ -1,7 +1,7 @@
 """The only object is exported - conversion.
 This object exposes the public API of conversions."""
-from .conversion import conversion  # noqa: F401
-from .dt import DateGrid, DateTimeGrid
+from ._conversion import conversion  # noqa: F401
+from ._dt import DateGrid, DateTimeGrid
 
 
-__version__ = "0.42.4"
+__version__ = "1.0.0"