+++ src/convtools/_cumulative.py
@@ -2,7 +2,7 @@
 import typing as t
 from uuid import uuid4
 
-from .base import (
+from ._base import (
     BaseConversion,
     EscapedString,
     If,